Move the transformations to elasticsearch_transformations file
